move basic tank (#13)
* tu16: checked out new branch from master - merge conflicts too complex to resolve confidently

* edited position struct to take in vector

* tu16: wip - todo rendertank next

* tu16: basic upfacing tank rendered

* added drawRectAdvanced

* matherbich i forgot semicolon

* tu16: wip (forgot what i did) destroy component function to cleanup tank components, tank can move

* revert change made by FOO BAR BOON last week

* tu16: dynamic keybinding and tank movement

* tu16: rotation done

* tu16: tank rotates fine

* reorg code

* tu16: resolved issue with different speeds when turning left/right

* tu16: reverted a lot of stuff man

* tu16: basic tank center(without moving

* tu16: wip - going to test so committing

* radius is correct

* wip - still working on rotating the goddamn rect on the center point

* wip - tank rotation. refer to math workings on shuting ipad

* wip - somewhat rotates, but still buggy

* wip - might try stupid shit soon so best commit first lol

* tu16: specified that rotation matrix used is clockwise

* tu16: converted some to radians

* backup

* tu16: separated get rect center logic

* realised i forgot to updaate tank

* made a bit of progress with issue finding

* tu116: i rly cry bro there was a setting for this this whole time?

* tu16: refining movement

* tu18: tank moves

* tu18: tank can accelerate and decelerate

* converted to doubles

+++ collision.c
@@ -1,6 +1,6 @@
 #include "config.h"
 
 
-bool mouseInRect(Rect r, float mouseX, float mouseY) {
+bool mouseInRect(Rect r, double mouseX, double mouseY) {
 	return (mouseX >= r.pos.x && mouseX <= r.pos.x + r.size.width && mouseY >= r.pos.y && mouseY <= r.pos.y + r.size.height);
 }
